:recycle: Refatoração Services
