+++ frontend/src/pages/Login/index.tsx
@@ -2,8 +2,7 @@ import React, { useState } from "react";
 import { useForm } from "react-hook-form";
 import * as Styled from "./styled";
 import { useNavigate } from "react-router";
-//import RotateBanner from '../../components/RotateBanner';
-import { login } from '../../service';
+import { login } from '../../service/login';
 import RotateBanner from '../../components/RotateBanner';
 import { FaEye, FaEyeSlash } from 'react-icons/fa';
 import logo from '../../images/agronomia.svg';
